remove CPU logging (moving to external script instead)
