+++ app-tracking-protection/vpn-impl/src/main/java/com/duckduckgo/mobile/android/vpn/health/AppTPCPUMonitor.kt
@@ -16,12 +16,10 @@
 
 package com.duckduckgo.mobile.android.vpn.health
 
-import android.content.Context
 import androidx.annotation.VisibleForTesting
 import androidx.work.ExistingPeriodicWorkPolicy
 import androidx.work.PeriodicWorkRequestBuilder
 import androidx.work.WorkManager
-import com.duckduckgo.appbuildconfig.api.AppBuildConfig
 import com.duckduckgo.di.scopes.VpnScope
 import com.duckduckgo.mobile.android.vpn.feature.AppTpFeatureConfig
 import com.duckduckgo.mobile.android.vpn.feature.AppTpSetting
@@ -37,9 +35,6 @@ import javax.inject.Inject
 class AppTPCPUMonitor @Inject constructor(
     private val workManager: WorkManager,
     private val appTpFeatureConfig: AppTpFeatureConfig,
-    private val appBuildConfig: AppBuildConfig,
-    private val context: Context,
-    private val cpuPerformanceLogger: CPUPerformanceLogger
 ) : VpnServiceCallbacks {
 
     companion object {
@@ -57,10 +52,6 @@ class AppTPCPUMonitor @Inject constructor(
         } else {
             Timber.d("AppTpSetting.CPUMonitoring is disabled")
         }
-
-        if (appBuildConfig.isPerformanceTest) {
-            cpuPerformanceLogger.startLogging(coroutineScope)
-        }
     }
 
     override fun onVpnStopped(
@@ -69,9 +60,5 @@ class AppTPCPUMonitor @Inject constructor(
     ) {
         Timber.v("AppTpSetting.CPUMonitoring - stopping")
         workManager.cancelUniqueWork(APP_TRACKER_CPU_MONITOR_WORKER_TAG)
-
-        if (appBuildConfig.isPerformanceTest) {
-            cpuPerformanceLogger.stopLogging(context)
-        }
     }
 }
