--- app-tracking-protection/vpn-impl/src/main/java/com/duckduckgo/mobile/android/vpn/health/CPUPerformanceLogger.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2022 DuckDuckGo
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.duckduckgo.mobile.android.vpn.health
-
-import android.content.Context
-import kotlinx.coroutines.CoroutineScope
-
-/**
- * Used during performance tests - logs CPU usage to a file
- */
-interface CPUPerformanceLogger {
-    fun startLogging(coroutineScope: CoroutineScope)
-    fun stopLogging(context: Context)
-}
